+++ apps/rosen-service2/src/app.ts
@@ -1,6 +1,9 @@
 import { DefaultLogger } from '@rosen-bridge/abstract-logger';
 import { ServiceManager } from '@rosen-bridge/service-manager';
 import { AssetAggregatorService } from 'services/assetAggregator';
+import { EventCountMetricService } from 'services/eventCountMetric';
+import { GeneralMetricsService } from 'services/generalMetrics';
+import { LockedAssetsMetricService } from 'services/lockedAssetsMetric';
 import { UserEventsMetricService } from 'services/userEventsMetric';
 
 import dataSource from './data-source';
@@ -35,12 +38,12 @@ const startApp = async () => {
   serviceManager.register(ScannerService.getInstance());
   logger.debug('Scanner service registered to the service manager');
 
-  logger.debug('Initializing user events metrics service');
-  UserEventsMetricService.init(
-    DefaultLogger.getInstance().child('userEventsMetricService'),
+  logger.debug('Initializing general metrics service');
+  GeneralMetricsService.init(
+    DefaultLogger.getInstance().child('generalMetricsService'),
   );
-  serviceManager.register(UserEventsMetricService.getInstance());
-  logger.debug('User events metrics service registered to the service manager');
+  serviceManager.register(GeneralMetricsService.getInstance());
+  logger.debug('General metrics service registered to the service manager');
 
   logger.debug('Initializing asset-data-adapter service');
   AssetDataAdapterService.init(
@@ -55,6 +58,29 @@ const startApp = async () => {
   );
   serviceManager.register(AssetAggregatorService.getInstance());
   logger.debug('asset-aggregator Service registered to the service manager');
+
+  logger.debug('Initializing locked assets metrics service');
+  LockedAssetsMetricService.init(
+    DefaultLogger.getInstance().child('lockedAssetsMetricsService'),
+  );
+  serviceManager.register(LockedAssetsMetricService.getInstance());
+  logger.debug(
+    'Locked assets metrics service registered to the service manager',
+  );
+
+  logger.debug('Initializing event count metrics service');
+  EventCountMetricService.init(
+    DefaultLogger.getInstance().child('eventCountMetricsService'),
+  );
+  serviceManager.register(EventCountMetricService.getInstance());
+  logger.debug('Event count metrics service registered to the service manager');
+
+  logger.debug('Initializing user events metrics service');
+  UserEventsMetricService.init(
+    DefaultLogger.getInstance().child('userEventsMetricService'),
+  );
+  serviceManager.register(UserEventsMetricService.getInstance());
+  logger.debug('User events metrics service registered to the service manager');
 
   logger.debug('Initializing health-check service');
   HealthService.init(DefaultLogger.getInstance().child('healthCheckService'));