+++ packages/ui-kit/src/utils/index.ts
@@ -1,3 +1,5 @@
-export * from './themeUtils';
 export * from './cx';
+export * from './calculateRelativeTime';
+export * from './formatDuration';
+export * from './themeUtils';
 export * from './themeOptions';