+++ networks/bitcoin-runes/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rosen-network/bitcoin-runes
 
+## 1.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/icons@3.5.0
+
 ## 1.0.3
 
 ### Patch Changes